move ReplicationBuffer to separate module
